Added network reduction by only sending position and rotation every 0.5 metres and/or 5 degrees
Added addition lines to player_syncPosition and player_syncRotation to record last known rotation and position.
upon rotating > 5 degrees and/or moving 0.5 metres then the players position and/or rotation is sent to the server. - last known position and or rotation are updated at this point.

this should reduce the amount of synching required across the network. Lerping also assists in smoothing rotation and movement