Historic Lerping added Added the ability to turn on or off historic lerping in the event of latency
The historic lerping is an list of vector3 - waypoints that the player has moved through. These are sent in retrospec to other clients so on their scene they see the player move through way points with a ms delay. almost like a replay - this provides a smoother movement on the screen. however the effects are unknown when using shooting and detection.

Lerping of rotation has as yet not been added